Get official data about Covid-19 cases in Mexico

Get official data about Covid-19 cases in Mexico. Main data source is Mexico's Federal Government Ministry of Health (Secretaría de Salud). It should be noted, however, that the open data format of the daily report is generated by third parties, since the original report is a PDF document. Wrapper for GetFromSerendipia, and GetFromGuzmart

getData(where = "Mexico", type = "confirmed", date = "today",
  source = "Guzmart", neat = TRUE)

Arguments

where
Scope of the report (Mexico or world wide). `character` vector of length 1.

type
Get `confirmed` or `suspect` cases?. `character` vector of length 1.

date
Date (version) of the published results. `character` vector of length 1 or `date` object.

source
Data source. Mexico's Ministry of Health publishes a daily report containing data about positive and suspected cases of Covid-19 cases in the country. However, the data is published as a PDF document, difficulting further analysis. This package relies on third party sources to get an "open data" version of the report. `character` vector of length 1. As of version 0.2.0, possible data sources are: `Serendipia` (confirmed, suspect), and `Guzmart` (confimed).

neat
Should data be cleaned (dates, state name, column names)? `logical` vector of length 1.

Value

A `tibble` with 8/9 columns: Case Number, State, Sex, Age, Symptoms Onset Date, COVID-19 Test Result, Country Visited, Date of Arrival to Mexico, and date when the case officialy registered (only available if `source = "Guzmart"`).